+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## Version 120001.01
+
+### Loot Window
+
+- Added Transmog Loot Helper and Can I Mog It? support.
+- While not in combat you'll be able to use the interaction hotkey and/or space to loot everything.
+
 ## Version 120000.02
 
 - Fixed various secret- and combat-related issues in the Adventure Guide.